fix: table status grid layout seems buggy
Fixes #4143
